+++ change-order/change-order-dao/src/main/java/me/cathub/change/order/dao/OrderItemDaoImpl.java
@@ -84,7 +84,7 @@ public class OrderItemDaoImpl extends BaseCrudMyBatisImpl<OrderItem> implements 
     }
 
     @Override
-    public long countByOrderId(long order_id, int tableIndex) throws Exception {
+    public int countByOrderId(long order_id, int tableIndex) throws Exception {
         int count = 0;
         try {
             HashMap<String, ? super Number> map = new HashMap<>();
